Refactor: sign 페이지 레이아웃 수정
