+++ packages/web/src/app/@modal/(.)sign-up/page.tsx
@@ -1,6 +1,7 @@
-import SignLayout from '@/containers/sign/SignLayout'
+import { BotIcon } from '@packages/ui'
 
 import Modal from '@/components/modal/modal'
+import SignLayout from '@/components/sign/SignLayout'
 
 export default function Page() {
   return (
@@ -15,6 +16,9 @@ export default function Page() {
             <p className="text-base not-italic font-bold leading-[170%] text-[#7d12ff]">
               도움이 필요하신가요?
             </p>
+            <i className="flex items-center justify-center pl-1">
+              <BotIcon />
+            </i>
           </div>
         </SignLayout.SignBox>
       </SignLayout>
