+++ packages/web/src/app/sign-up/page.tsx
@@ -1,4 +1,6 @@
-import SignLayout from '@/containers/sign/SignLayout'
+import { BotIcon } from '@packages/ui'
+
+import SignLayout from '@/components/sign/SignLayout'
 
 export default function Page() {
   return (
@@ -14,6 +16,9 @@ export default function Page() {
             <p className="text-base not-italic font-normal leading-[170%] text-[#7d12ff]">
               도움이 필요하신가요?
             </p>
+            <i className="flex items-center justify-center pl-1">
+              <BotIcon />
+            </i>
           </div>
         </SignLayout.SignBox>
       </SignLayout>
